app/backend/requirements.txt
flask
flask-cors
Pillow
redis
celery
numpy
tensorflow